Add DDS24 files under components folder.
Previously the DDS24 files were outside the project
directory. Therefore everytime the project is downloaded from the remote
repository, the dependency needs to be added. Making this change will
remove the need to add in the dependancy everytime the repository is
cloned since the DDS24 files will be inside the project and the path
will remain the same.